style: Move grid-template-areas outside of mako
